refactor: move the operator tests next to the modules they exercise
The 766-line operator/tests.rs sat at a re-export-only parent and covered
both children. Split along the design/schwarz seam; the design_of fixture
becomes Design::from_levels_for_test, matching CrossTab::from_dense_for_test.
